# apps/web/reports/api-spec.yaml
openapi: 3.1.0
info:
  title: ApexGraphSwarm API
  description: |
    ApexGraphSwarm workbench server routes. All endpoints return JSON with
    `Cache-Control: no-store`. Authentication is via bearer tokens validated
    with constant-time comparison.
    
    Three distinct authentication schemes exist across routes:
    - **Integration token** (`INTEGRATION_ACCESS_TOKEN`): used by analytics, control,
      decisions (POST), ecosystem/mcp (POST), integrations, integrations/jobs/[id],
      optimization, and planned-tasks.
    - **Graph store token** (`GRAPH_STORE_ACCESS_TOKEN`): used by graph-store.
    - **Model review token** (`GRAPH_REVIEW_ACCESS_TOKEN`): used by review and swarm.
    
    GET-only routes (decisions, ecosystem/mcp, graph-store, integrations, review)
    are unauthenticated.
    
    Origin validation (`hasIntegrationSafeOrigin`, `hasGraphStoreSafeOrigin`,
    `hasSafeOrigin`) checks the Origin header matches the Host; missing Origin
    is treated as a non-browser caller and allowed.
  version: 0.1.0
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - integrationBearer: []
paths:
...
  /api/graph-store:
    get:
      tags: [graph-store]
      summary: Check graph store status
      description: Returns whether Neo4j graph storage is enabled and configured.
      operationId: getGraphStoreStatus
      security: []
      responses:
        '200':
          description: Graph store status
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
                  mode:
                    type: string
                    enum: [neo4j]
    post:
      tags: [graph-store]
      summary: Load or save a graph snapshot
      description: |
        Loads the active graph revision or saves a new one to Neo4j. Request body
        limited to 2 MiB. Graph limited to 20,000 nodes and 80,000 edges.
        Times out after 20 seconds.
      operationId: graphStoreOperation
      security:
          graphStoreBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [load, save]
                graph:
                  $ref: '#/components/schemas/Snapshot'
                  description: Required when action is "save"
      responses:
        '200':
          description: Graph snapshot (load) or stored revision metadata (save)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Snapshot'
                  - $ref: '#/components/schemas/StoredRevision'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Request origin not allowed
        '404':
          description: No saved graph available
        '413':
          description: Request body exceeds 2 MiB
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '502':
          description: Graph storage failed
        '503':
          description: Neo4j not configured
        '504':
          description: Graph storage request timed out
...
components:
  securitySchemes:
    integrationBearer:
      type: http
      scheme: bearer
      description: INTEGRATION_ACCESS_TOKEN bearer token
    graphStoreBearer:
      type: http
      scheme: bearer
      description: GRAPH_STORE_ACCESS_TOKEN bearer token
    reviewBearer:
      type: http
      scheme: bearer
      description: GRAPH_REVIEW_ACCESS_TOKEN bearer token
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    UnsupportedMediaType:
      description: Content-Type must be application/json
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
    Snapshot:
      type: object
      required: [version, name, nodes, edges]
      properties:
        version:
          type: integer
          minimum: 1
        name:
          type: string
          maxLength: 500
        nodes:
          type: array
          maxItems: 25000
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          maxItems: 100000
          items:
            $ref: '#/components/schemas/GraphEdge'
        warnings:
          type: array
          maxItems: 1000
          items:
            type: string
        truncated:
          type: boolean
        summary:
          type: object
        unresolved:
          type: array
          maxItems: 200
          items:
            type: object
            properties:
              path: { type: string }
              line: { type: integer }
              expression: { type: string }
    GraphNode:
      type: object
      required: [id, name, kind, path, confidence]
      properties:
        id:
          type: string
          maxLength: 2000
        name:
          type: string
          maxLength: 2000
        kind:
          type: string
          enum: [module, file, function, class, external]
        path:
          type: string
          maxLength: 4000
        summary:
          type: string
          maxLength: 10000
        confidence:
          type: string
          enum: [parsed, observed, inferred, illustrative, aggregated]
        line:
          type: integer
          minimum: 1
        connections:
          type: integer
          minimum: 0
    GraphEdge:
      type: object
      required: [source, target, relation, confidence]
      properties:
        source:
          type: string
        target:
          type: string
        relation:
          type: string
          maxLength: 100
        confidence:
          type: string
          enum: [parsed, observed, inferred, illustrative, aggregated]
        line:
          type: integer
        count:
          type: integer
          minimum: 1
    StoredRevision:
      type: object
      properties:
        revision:
          type: string
        name:
          type: string
        nodeCount:
          type: integer
        edgeCount:
          type: integer
        createdAt:
          type: string
          nullable: true
...